refactor(editor): Move workflowHistory api to rest-api-client
