+++ packages/frontend/editor-ui/src/stores/__tests__/utils/workflowHistoryTestUtils.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import type { WorkflowHistory, WorkflowVersion } from '@/types/workflowHistory';
+import type { WorkflowHistory, WorkflowVersion } from '@n8n/rest-api-client/api/workflowHistory';
 
 export const workflowHistoryDataFactory: () => WorkflowHistory = () => ({
 	versionId: faker.string.nanoid(),
