+++ packages/frontend/editor-ui/src/views/WorkflowHistory.test.ts
@@ -17,7 +17,7 @@ import {
 	workflowHistoryDataFactory,
 	workflowVersionDataFactory,
 } from '@/stores/__tests__/utils/workflowHistoryTestUtils';
-import type { WorkflowVersion } from '@/types/workflowHistory';
+import type { WorkflowVersion } from '@n8n/rest-api-client/api/workflowHistory';
 import type { IWorkflowDb } from '@/Interface';
 import { telemetry } from '@/plugins/telemetry';
